requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
automat==25.4.16
cffi==2.0.0
channels==4.3.1
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
django==4.2.25
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyopenssl==25.3.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.3.0
twisted==25.5.0
txaio==23.6.1
typing-extensions==4.15.0
tzdata==2025.2
zope-interface==8.0.1
